New image saving logic
Implemented a dictionary to keep track of detected images and their confidence level.

When an image class is detected, it will be saved regardless of confidence level. But if the same image class is detected again with a higher confidence level, the older image will be overwritten and the higher confidence image will be saved.